Fix Stand/Crouch/Jump signature mismatches (int->short) for ChunLi, Zangeif, Dhalsim
Same class of WASM freeze as PLCBPower — function pointer arrays typed as
short(*)(Player*) but implementations returned int. Removed all forced casts.
Also adds charselect mask with hole for 1P cursor area.

+++ FistBlue/avatars/chunli.h
@@ -25,9 +25,9 @@ void pl_cb_setstatus2_chunli(Player *player, short status, int argd0);
 void pl_cb_setstatus3_chunli(Player *player, short status);
 
 void PLCBPowerChunLi(Player *ply);
-int PLCBStandChunLi(Player *ply);
-int PLCBCrouchChunLi(Player *ply);
-int PLCBJumpChunLi(Player *ply);
+short PLCBStandChunLi(Player *ply);
+short PLCBCrouchChunLi(Player *ply);
+short PLCBJumpChunLi(Player *ply);
 void PSCBAttackChunLi(Player *ply);
 
 void PLCBCompAttackChunLi(Player *ply);
